added reusable decorators for stories
