+++ apps/ui-sharethrift/tsconfig.json
@@ -11,6 +11,6 @@
 		"moduleResolution": "bundler",
 		"allowImportingTsExtensions": true
 	},
-	"include": ["src"],
+	"include": ["src/**/*"],
 	"exclude": ["node_modules", "dist"]
 }
